requirements/base.txt
cmsplugin-filer==1.1.2
Django==1.9.9
django-appconf==1.0.2
django-classy-tags==0.7.2
django-cms==3.3.2
django-debug-toolbar==1.5
django-environ==0.4.0
django-filer==1.2.4
django-formtools==1.0
django-mptt==0.8.5
django-polymorphic==0.8.1
django-reversion==2.0.0
django-sekizai==0.9.0
Django-Select2==5.8.9
django-treebeard==4.0.1
djangocms-admin-style==1.2.3
djangocms-attributes-field==0.1.1
djangocms-column==1.6.0
djangocms-file==1.0
djangocms-googlemap==0.5.1
djangocms-inherit==0.2.1
djangocms-link==1.8.2
djangocms-picture==1.0.0
djangocms-style==1.7.0
djangocms-teaser==0.2.0
djangocms-text-ckeditor==2.9.5
djangocms-video==1.1.0
easy-thumbnails==2.3
html5lib==0.9999999
Pillow==3.3.1
psycopg2==2.6.2
six==1.10.0
sqlparse==0.2.1
Unidecode==0.4.19
whitenoise==3.2.1